refactor(shared): move formatWorldBytes to dependency-free utils/format module
domain 엔티티(World)가 node:fs를 끌어오는 utils 배럴 대신 순수 모듈을 import하도록 분리(Clean Architecture).
